integration test for InResponseTo error
Creates a new integration test, which checks the handling of errors in
various cases of in_response_to header vs saml token metadata mismatch.
Extracts common test base for saml IT tests from the existing
SamlServiceProviderMetadataIT
